move related input processing into relevant module
